Added Bloc state management Added global AppText and AppDimensions
Added Todo List view with infinite scroll and Details view
Added repository for retrieving todos and definitions

Added support for when the app is closed and reopened - using Hive as a local db

Added additional helpers, Date Formatter, Colour based on definition

Added Api Client - this should be moved out of the Features section as it will be a globally used thing.

Added different templates to accomodate for DEADLINE, SIMPLE, CHECKBOX and using a factory which expects the type definition it'll render the relevant screen.